+++ src/selection.h
@@ -5,14 +5,14 @@
 
 typedef struct {
     BlockIter si;
-    long so;
-    long eo;
+    size_t so;
+    size_t eo;
     bool swapped;
 } SelectionInfo;
 
 void init_selection(const View *v, SelectionInfo *info);
-long prepare_selection(View *v);
-char *view_get_selection(View *v, long *size);
+size_t prepare_selection(View *v);
+char *view_get_selection(View *v, size_t *size);
 size_t get_nr_selected_lines(const SelectionInfo *info);
 size_t get_nr_selected_chars(const SelectionInfo *info);
